Warn user if client certificate is expired
If a user's client certificate is expired, the API server will reject a
kubectl request with a vague 401 error. Adding more information to the
server's response risks exposing too much information to a potential
attacker, but it's reasonable to add client-side validation prior making
the request. This change to client-go parses the user's certificate upon
loading it and checks whether it is valid according to the host's system
time. If the certificate is expired, it logs a warning but allows the
request to continue; this is to accomodate the case that the user's host
system may not have the correct time set but the certificate is still
valid to the API server, while still leaving an appropriate hint to the
user.

Since the unit tests for the transport package were already using an
expired certificate, this change adds a new certificate that will not
expire for 100 years to avoid superfluous log warnings in the test
output. A more robust solution might be to generate a new certificate on
each test run but this would add some time to the test run and some
extra complexity to the test setup.